requirements.txt
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1 